Fix lint-y issues, including a questionable race
The async/await one in particular I think meant those fetches weren't
actually being awaited, which would cause the user creation to race with
adding the users to the group. If the add-to-group won the race, it
would error out, and not set up the app correctly -- but be fine on
refresh!